Fix TimestampColumnWriter NoSuchMethodError
Java 11 has 3 floor mods. (long, long), (long, int) and (int, int).
https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/lang/Math.html#floorMod(long,int)

java 8 has 2 floor mods (long, long) and (long, int).
https://docs.oracle.com/javase/8/docs/api/java/lang/Math.html#floorMod-long-long-

The code is most likely (my best guess from previous experiences) is compiled with java11 binary
but compatibility is set to java8. It is linking against the java (long,int) and returning int
overload. But java8 does not have it when we run with java8 JRE and failing with it.

Caused by: java.lang.NoSuchMethodError: java.lang.Math.floorMod(JI)I

+++ presto-orc/src/main/java/com/facebook/presto/orc/writer/TimestampColumnWriter.java
@@ -82,7 +82,7 @@ public class TimestampColumnWriter
     private long columnStatisticsRetainedSizeInBytes;
 
     private final long baseTimestampInSeconds;
-    private final int unitsPerSecond;
+    private final long unitsPerSecond;
     private final int trailingZeros;
 
     private int nonNullValueCount;
